# apps/luna-shopper-backend/gateway/docs/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/zones": {
      "post": {
        "operationId": "ZoneController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateZoneDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The new zone. `tokens` is present when the caller was anonymous.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/zone.ZoneView.WithMaybeToken"
                }
              }
            }
          },
          "400": {
            "description": "`validation_failed` — Some of the submitted values are invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` — Too many requests. Please slow down and try again shortly.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "`internal` — Something went wrong on our side. Please try again.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "tags": [
          "zones"
        ]
      },
      "get": {
        "operationId": "ZoneController_listMine_v1",
        "parameters": [
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "name",
                "joined",
                "recent"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The `zone.listMine` contract response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/zone.ZonePage"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized` — You need to sign in to do that.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` — Too many requests. Please slow down and try again shortly.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "`internal` — Something went wrong on our side. Please try again.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "zones"
        ]
      }
    },
    "/v1/zones/join": {
      "post": {
        "operationId": "ZoneController_join_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinZoneDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The pending membership. `tokens` is present when the caller was anonymous.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/zone.MembershipView.WithMaybeToken"
                }
              }
            }
          },
          "400": {
            "description": "`validation_failed` — Some of the submitted values are invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "`forbidden` — You do not have permission to do that.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "`not_found` — The requested resource was not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "`conflict` — That request conflicts with the current state.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` — Too many requests. Please slow down and try again shortly.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "`internal` — Something went wrong on our side. Please try again.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "tags": [
          "zones"
        ]
      }
    },
    "/v1/zones/{id}": {
      "patch": {
        "operationId": "ZoneController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateZoneDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The `zone.update` contract response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/zone.ZoneView"
                }
              }
            }
          },
          "400": {
            "description": "`validation_failed` — Some of the submitted values are invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized` — You need to sign in to do that.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "`forbidden` — You do not have permission to do that.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "`not_found` — The requested resource was not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` — Too many requests. Please slow down and try again shortly.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "`internal` — Something went wrong on our side. Please try again.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "zones"
        ]
      },
      "delete": {
        "operationId": "ZoneController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The `zone.delete` contract response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/common.IdResult"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized` — You need to sign in to do that.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "`forbidden` — You do not have permission to do that.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "`not_found` — The requested resource was not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` — Too many requests. Please slow down and try again shortly.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "`internal` — Something went wrong on our side. Please try again.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "zones"
        ]
      }
    },
    "/v1/zones/{id}/regenerate-code": {
      "post": {
        "operationId": "ZoneController_regenerateCode_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "The `zone.regenerateJoinCode` contract response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/zone.ZoneView"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized` — You need to sign in to do that.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "`forbidden` — You do not have permission to do that.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "`not_found` — The requested resource was not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` — Too many requests. Please slow down and try again shortly.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "`internal` — Something went wrong on our side. Please try again.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "zones"
        ]
      }
    },
    "/v1/zones/{id}/claim-ownership": {
      "post": {
        "operationId": "ZoneController_claimOwnership_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "The `zone.claimOwnership` contract response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/zone.ZoneView"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized` — You need to sign in to do that.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "`forbidden` — You do not have permission to do that.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "`not_found` — The requested resource was not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "`conflict` — That request conflicts with the current state.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` — Too many requests. Please slow down and try again shortly.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "`internal` — Something went wrong on our side. Please try again.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "zones"
        ]
      }
    },
    "/v1/zones/{id}/members/{membershipId}/role": {
      "patch": {
        "operationId": "ZoneController_setRole_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "membershipId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The `zone.setRole` contract response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/zone.MembershipView"
                }
              }
            }
          },
          "400": {
            "description": "`validation_failed` — Some of the submitted values are invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized` — You need to sign in to do that.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "`forbidden` — You do not have permission to do that.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "`not_found` — The requested resource was not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` — Too many requests. Please slow down and try again shortly.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "`internal` — Something went wrong on our side. Please try again.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "zones"
        ]
      }
    },
...
  },
  "info": {
    "title": "Luna Shopper API",
    "description": "Public API for the Luna Shopper shared shopping lists. Response schemas are projected from the same JSON Schemas the services validate their broker messages against, so what is documented here is what the wire carries.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateZoneDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 80
          },
          "username": {
            "type": "string",
            "maxLength": 40,
            "description": "Your display name in this zone"
          }
        },
        "required": [
          "name",
          "username"
        ]
      },
      "JoinZoneDto": {
        "type": "object",
        "properties": {
          "joinCode": {
            "type": "string"
          },
          "username": {
            "type": "string",
            "maxLength": 40,
            "description": "Your display name in this zone"
          }
        },
        "required": [
          "joinCode",
          "username"
        ]
      },
...
      "zone.ZonePage": {
        "type": "object",
        "description": "A cursor paginated page. `nextCursor` is null on the last page; otherwise pass it back as the `cursor` query parameter to fetch the next one.",
        "additionalProperties": false,
        "required": [
          "items",
          "nextCursor"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/zone.MyZoneView"
            }
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "zone.ZoneView.WithMaybeToken": {
        "type": "object",
        "description": "The result, plus the identity minted for the caller. `tokens` is present only when the request arrived without an `Authorization` header: the endpoint created a temporary user to own the operation, and these are the only credentials that will ever be issued for it. An authenticated caller gets `data` alone.",
        "required": [
          "data"
        ],
        "additionalProperties": false,
        "properties": {
          "tokens": {
            "$ref": "#/components/schemas/auth.AuthTokens",
            "description": "Present only when the caller was anonymous."
          },
          "data": {
            "$ref": "#/components/schemas/zone.ZoneView"
          }
        }
      }
    }
  }
}